requirements/requirements_dev.txt
asgiref==1.1.2
asgi-redis==1.4.3
channels==1.1.8
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
bcrypt==3.1.4
cffi==1.11.4
click==6.7
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.1.4
-e git+ssh://git@github.com/aeguana/digitalvendcore.git#egg=digitalvendcore
Django==1.11.5
django-appconf==1.0.2
django-compressor==2.1
django-debug-toolbar==1.9.1
django-libsass==0.7
django-taggit==0.21.3
djangorestframework==3.4.7
djangorestframework-camel-case==0.2.0
Fabric3==1.13.1.post1
honcho==0.7.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
jsonfield==1.0.3
libsass==0.11.1
paramiko==2.4.0
pika==0.10.0
psutil==5.4.3
pyasn1==0.4.2
pycparser==2.18
PyCRC==1.21
PyNaCl==1.2.1
pyserial==3.2.1
python-dotenv==0.7.1
pytz==2017.2
raven==6.1.0
rcssmin==1.0.6
redis==2.10.6
requests==2.12.3
rjsmin==1.0.12
setproctitle==1.1.10
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
zope.interface==4.4.3
